FIX-#26: Reduce memory consumption in `Python` backend
