--- docs/flow/unidist/core/backends/python/core/object_store.rst
@@ -1,17 +1,0 @@
-..
-      Copyright (C) 2021 Modin authors
-
-      SPDX-License-Identifier: Apache-2.0
-
-:orphan:
-
-Object Storage
-""""""""""""""
-
-Python :py:class:`~unidist.core.backends.python.core.object_store.ObjectStore` stores
-data in the `standard python dict`_.
-
-.. autoclass:: unidist.core.backends.python.core.object_store.ObjectStore
-  :members:
-
-.. _`standard python dict`: https://docs.python.org/3/tutorial/datastructures.html#dictionaries
