+++ unidist/core/backends/python/core/__init__.py
@@ -4,6 +4,6 @@
 
 """Python backend core functionality."""
 
-from .api import put, get, submit, init
+from .api import wrap, unwrap, submit
 
-__all__ = ["put", "get", "submit", "init"]
+__all__ = ["wrap", "unwrap", "submit"]
